requirements.txt
django==2.2
djangorestframework==3.9.2
djangorestframework-filters==0.10.2
djangorestframework-jwt==1.11.0
drf-tracking==1.5.0
django-filter==2.1.0
django-reversion==3.0.3
django-cors-headers==2.5.2
django-bootstrap4==0.0.8
bootstrap-admin==0.4.3
django-sendsms==0.3.1
django-anymail[mailgun]==6.0
django-phonenumber-field==2.3.1
phonenumbers==8.10.10
simplejson==3.16.0
psycopg2-binary==2.8.2
python-dateutil==2.8.0
requests==2.21.0
twilio==6.26.1
isodate==0.6.0
pillow==6.0.0
pytest==4.4.1
pytest-cov==2.6.1
pytest-django==3.4.8
